Init basic typescript config
