+++ frontend challenge/src/App.tsx
@@ -5,9 +5,7 @@ import Typography from "@material-ui/core/Typography";
 function App() {
   return (
     <Container maxWidth="md">
-      <Typography variant="h1">
-        Icon cards
-      </Typography>
+      <Typography variant="h1">Icon cards</Typography>
       Edit <code>src/App.js</code> and save to reload.
     </Container>
   );
